Move ErrorSink to Errors folder
